Move normalize_hostname to helpers to break circular import
Importing from sinks.socket in vulnerabilities.ssrf.imds created a
cycle through sinks.socket.__init__ → aikido_zen.vulnerabilities.
Move the implementation to helpers/normalize_hostname.py and have
sinks/socket/normalize_hostname.py re-export from there.
