+++ aikido_zen/sinks/socket/normalize_hostname.py
@@ -1,14 +1,3 @@
-def normalize_hostname(hostname):
-    if not hostname or not isinstance(hostname, str):
-        return hostname
+from aikido_zen.helpers.normalize_hostname import normalize_hostname
 
-    # Lowercase and strip trailing dot (DNS resolvers may return FQDNs like "example.com.")
-    result = hostname.lower().rstrip(".")
-
-    try:
-        # Decode Punycode if the hostname starts with xn--
-        if result.startswith("xn--"):
-            result = result.encode("ascii").decode("idna")
-        return result
-    except (UnicodeError, LookupError):
-        return result
+__all__ = ["normalize_hostname"]
