+++ aikido_zen/vulnerabilities/ssrf/imds.py
@@ -4,7 +4,7 @@ is_imds_ip_address, is_trusted_hostname, resolves_to_imds_ip
 """
 
 from aikido_zen.helpers.ip_matcher import IPMatcher
-from aikido_zen.sinks.socket.normalize_hostname import normalize_hostname
+from aikido_zen.helpers.normalize_hostname import normalize_hostname
 
 imds_addresses = IPMatcher(
     [
